[style]: Move Shuffler
